pyproject.toml
[tool.poetry]
name = "fluxr"
version = "0.1.0"
description = "Task Automation Framework"
authors = ["Igor Medeiros <igor.medeiros@gmail.com>"]
readme = "README.md"
packages = [
    {include = "components", from = "src"},
    {include = "core", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
numpy = "^2.1.3"
flask = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"